requirements.txt
streamlit
python-docx
PyMuPDF
msoffcrypto-tool
langdetect
requests
re
